15146 DRY the edge shadow modifier
